Core Functionality
Created a program that imports a set of static objects representing groceries and outputs them to a shopping catalogue.

The user can add and remove these items to and from a checkout list.

The user can also filter the data with a dropdown box based on the grocery's category.

The page is also responsive and compatible with mobile devices.